+++ src/main/java/br/com/server/repository/ContatoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.server.model.Contato;
 
 @Repository
-public interface ContatoRepository extends JpaRepository<Contato, Integer> {
+public interface ContatoRepository extends JpaRepository<Contato, Long> {
 
 }
